integrate client into build
