+++ buildSrc/src/main/groovy/uk/gov/hmcts/AngularExec.groovy
@@ -4,6 +4,7 @@ import org.gradle.api.tasks.Exec
 
 class AngularExec extends Exec {
     AngularExec() {
+        dependsOn('build')
         getInputs().files(project.fileTree(project.projectDir).matching {
             include "*.json"
         });
