+++ java/build.gradle
@@ -4,7 +4,6 @@ plugins {
     id 'com.revolut.jooq-docker' version '0.3.5'
     id 'groovy'
     id 'uk.gov.hmcts.java' version '0.12.2'
-    id 'org.openapi.generator' version '4.3.1'
 }
 
 group = 'uk.gov.hmcts.ccf'
@@ -100,8 +99,3 @@ tasks.withType(Checkstyle) {
     exclude '**/org/jooq/**'
 }
 
-task buildTypescriptClient(type: org.openapitools.generator.gradle.plugin.tasks.GenerateTask, dependsOn: integrationTest) {
-    generatorName = "typescript-angular"
-    inputSpec = file("$buildDir/open-api.yaml").toString()
-    outputDir = "$buildDir/../../angular/src/generated/client-lib".toString()
-}
